// src/api/model/extraction-status.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */

export type ExtractionStatus =
	(typeof ExtractionStatus)[keyof typeof ExtractionStatus]

export const ExtractionStatus = {
	PENDING: "PENDING",
	IN_PROGRESS: "IN_PROGRESS",
	BACKEND_ERROR: "BACKEND_ERROR",
	AI_SERVICE_ERROR: "AI_SERVICE_ERROR",
	FINISHED_WITH_TASKS: "FINISHED_WITH_TASKS",
	FINISHED_WITHOUT_TASKS: "FINISHED_WITHOUT_TASKS",
} as const